+++ backend/src/main/java/dk/opendesk/repo/beans/AuthorityBean.java
@@ -1,10 +1,10 @@
-// 
+//
 // Copyright (c) 2017-2018, Magenta ApS
-// 
+//
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
-// 
+//
 
 package dk.opendesk.repo.beans;
 
@@ -78,6 +78,10 @@ public class AuthorityBean {
         json.put("fullName", fullName);
         String displayName = authorityService.getAuthorityDisplayName(fullName);
         json.put("displayName", displayName);
+        if (shortName.startsWith("BSK_"))
+            shortName = shortName.substring(4);
+        String displayFullName = displayName + " (" + shortName + ")";
+        json.put("displayFullName", displayFullName);
         String avatar = "assets/img/avatars/blank-profile-picture.png";
         json.put("avatar", avatar);
         return json;