+++ frontend/app/src/backendConfig.module.js
@@ -1,16 +1,17 @@
-// 
+//
 // Copyright (c) 2017-2018, Magenta ApS
-// 
+//
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
-// 
+//
 
 'use strict'
 
 angular
   .module('openDeskApp.backendConfig', [])
   .constant('APP_BACKEND_CONFIG', {
+    enableFavouriteSiteNotifications: false,
     enableProjects: false,
     enableSites: true,
     enableESDH: false,