+++ NEWS.md
@@ -6,3 +6,20 @@
 2. Please use the Markdown syntax.
 
 * Add feature descriptions here...
+
+Feature
+
+* Settings are now stored in alfresco's memory and are updated
+  when the settings node is updated (from the update webscript).
+  
+* Added configuration for notification limit to favourite sites/projects. (Redmine #28817)
+
+* Added admin webscript (/reloadEditors) to reload editor list. (Redmine #29976)
+
+* Security updates of frontend development packages (Redmine #29965)
+
+Bugfixes
+
+* Fixed problem with notification changes (read and seen) being shown on My Documents list in Share (Redmine #28186) 
+
+* Fixed problem with folders not displaying their content when one or more files are missing mime types. (Redmine #29683)